backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
cryptography==42.0.5
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastapi==0.110.0
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
idna==3.6
ipykernel==6.27.1
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
parso==0.8.3
passlib==1.7.4
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-multipart==0.0.9
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.12
Werkzeug==3.0.1
yarl==1.9.4